requirements.txt
xlsxwriter
Pillow
pywin32
psutil